requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aenum==3.1.15
    # via line-bot-sdk
aiohttp==3.9.3
    # via line-bot-sdk
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
deprecated==1.2.14
    # via line-bot-sdk
fastapi==0.110.1
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
future==1.0.0
    # via line-bot-sdk
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via
    #   anyio
    #   requests
    #   yarl
line-bot-sdk==3.10.1
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pydantic==2.7.1
    # via
    #   fastapi
    #   line-bot-sdk
pydantic-core==2.18.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via line-bot-sdk
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   uvicorn
requests==2.31.0
    # via line-bot-sdk
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   line-bot-sdk
    #   requests
uvicorn==0.29.0
    # via -r requirements.in
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via aiohttp
